pyproject.toml
[tool.poetry]
name = "devision-business-logic-fastapi"
version = "0.1.0"
description = ""
authors = ["YASHALAWA8 <95289183+YASHALAWA8@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.0"}
sqlalchemy = {extras = ["all"], version = "^2.0.35"}
alembic = "^1.13.3"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pydantic = "^2.9.2"
ruff = "^0.6.9"
mypy = "^1.11.2"
make = "^0.1.6.post2"
coverage = "^7.6.1"
aiofile = "^3.9.0"
aiofiles = "^24.1.0"
sqladmin = {extras = ["full"], version = "^0.20.0"}
pyjwt = "^2.9.0"


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]

select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"

docstring-code-format = false

docstring-code-line-length = "dynamic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["apps/", "tests/"]    # Папка для проверки типов
ignore_missing_imports = true 